backend/requirements.txt
flask
flask-cors
pyttsx3
sounddevice
soundfile
whisper
language_tool_python
praat-parselmouth
numpy
openai
faster-whisper
deepgram-sdk==2.12.0
python-dotenv
watchdog
PyJWT
# System dependencies required:
# - ffmpeg (for audio conversion)
# - Praat (for parselmouth, if not included) 